# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minizon API",
        "description": "API REST de la plateforme Minizon — Transport et livraison.",
        "contact": {
            "name": "Équipe Minizon",
            "email": "dev@minizon.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/vehicles": {
            "get": {
                "tags": [
                    "🚙 Flotte & Véhicules"
                ],
                "summary": "Lister les véhicules",
                "description": "Un **conducteur** ne voit que ses propres véhicules. Un **administrateur** accède à la flotte globale avec les informations du propriétaire.",
                "operationId": "vehiclesIndex",
                "responses": {
                    "200": {
                        "description": "Liste des véhicules récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des véhicules récupérée."
                                        },
                                        "body": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "🚙 Flotte & Véhicules"
                ],
                "summary": "Soumettre un nouveau véhicule",
                "description": "Permet à un conducteur connecté de soumettre un véhicule pour certification. Le véhicule est créé avec `is_approved = false` et doit être validé par un administrateur avant d'être utilisable.",
                "operationId": "vehiclesStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_type_id",
                                    "brand",
                                    "model",
                                    "color",
                                    "available_seats",
                                    "license_plate",
                                    "vehicle_photo",
                                    "registration_doc",
                                    "insurance_doc",
                                    "tvm_doc"
                                ],
                                "properties": {
                                    "vehicle_type_id": {
                                        "description": "ID du type de véhicule (voir table vehicle_types)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Toyota"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "Camry"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Gris Étoile"
                                    },
                                    "available_seats": {
                                        "description": "Nombre de places disponibles (hors conducteur)",
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "license_plate": {
                                        "description": "Immatriculation unique",
                                        "type": "string",
                                        "example": "RB 1234 X"
                                    },
                                    "vehicle_photo": {
                                        "description": "Chemin de la photo du véhicule",
                                        "type": "string",
                                        "example": "uploads/vehicles/photo.png"
                                    },
                                    "registration_doc": {
                                        "description": "Carte grise",
                                        "type": "string",
                                        "example": "uploads/docs/carte_grise.pdf"
                                    },
                                    "insurance_doc": {
                                        "description": "Attestation d'assurance",
                                        "type": "string",
                                        "example": "uploads/docs/assurance.pdf"
                                    },
                                    "tvm_doc": {
                                        "description": "TVM (Taxe sur les Véhicules à Moteur)",
                                        "type": "string",
                                        "example": "uploads/docs/tvm.pdf"
                                    },
                                    "technical_control_doc": {
                                        "description": "Visite technique (optionnel)",
                                        "type": "string",
                                        "example": "uploads/docs/ct.pdf",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Véhicule soumis — en attente de certification admin"
                    },
                    "422": {
                        "description": "Données invalides ou immatriculation déjà existante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/vehicles/{id}/approve": {
            "post": {
                "tags": [
                    "🚙 Flotte & Véhicules"
                ],
                "summary": "[ADMIN] Certifier / Suspendre un véhicule",
                "description": "Action réservée au Back-Office. Active ou gèle l'autorisation de rouler d'un véhicule (`is_approved`). Un véhicule non certifié ne peut pas être utilisé pour créer un trajet.",
                "operationId": "vehiclesToggleApproval",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant numérique du véhicule",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "is_approved"
                                ],
                                "properties": {
                                    "is_approved": {
                                        "description": "`true` pour certifier, `false` pour suspendre",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut de certification du véhicule mis à jour"
                    },
                    "403": {
                        "description": "Privilèges administratifs requis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Véhicule introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Valeur d'approbation invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Message d'erreur."
                    },
                    "body": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}